Suggested changes + cleanup
