+++ Samples/3.0/AspNetCoreCohosting/API/Controllers/HelloWorldController.cs
@@ -1,10 +1,9 @@
-﻿using System;
 using System.Threading.Tasks;
-using AspNetCoreHostedServices.Interfaces;
+using AspNetCoreCohosting.Interfaces;
 using Microsoft.AspNetCore.Mvc;
 using Orleans;
 
-namespace ASPNetCoreHostedServices.Controllers
+namespace AspNetCoreCohosting.Controllers
 {
     [ApiController]
     [Route("api/hello")]
@@ -13,9 +12,10 @@ namespace ASPNetCoreHostedServices.Controllers
         private readonly IClusterClient _client;
         private readonly IHelloWorld _grain;
 
-        public HelloWorldController(IClusterClient client) {
-            this._client = client;
-            this._grain = this._client.GetGrain<IHelloWorld>(0);
+        public HelloWorldController(IClusterClient client)
+        {
+            _client = client;
+            _grain = _client.GetGrain<IHelloWorld>(0);
         }
 
         [HttpGet]
