+++ Samples/3.0/AspNetCoreCohosting/API/Properties/launchSettings.json
