+++ Samples/3.0/AspNetCoreCohosting/API/appsettings.Development.json
