+++ Samples/3.0/AspNetCoreCohosting/Grains/Grains.csproj
