+++ Samples/3.0/AspNetCoreCohosting/Interfaces/Interfaces.csproj
