--- Samples/3.0/HostedService/API/API.csproj
@@ -1,19 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.0</TargetFramework>
-    <RootNamespace>ASPNetCoreHostedServices</RootNamespace>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Orleans.Core" Version="3.0.0" />
-    <PackageReference Include="Microsoft.Orleans.OrleansRuntime" Version="3.0.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Grains\Grains.csproj" />
-    <ProjectReference Include="..\Interfaces\Interfaces.csproj" />
-    <ProjectReference Include="..\Silo\Silo.csproj" />
-  </ItemGroup>
-
-</Project>
