--- Samples/3.0/HostedService/API/Startup.cs
@@ -1,50 +1,0 @@
-using AspNetCoreHostedServices.Grains;
-using AspNetCoreHostedServices.Silo;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-
-namespace ASPNetCoreHostedServices
-{
-    public class Startup
-    {
-        public Startup(IConfiguration configuration)
-        {
-            Configuration = configuration;
-        }
-
-        public IConfiguration Configuration { get; }
-
-        // This method gets called by the runtime. Use this method to add services to the container.
-        public void ConfigureServices(IServiceCollection services)
-        {
-            services.AddHostedService<HelloWorldSiloHostedService>();
-            services.AddSingleton<HelloWorldClientHostedService>();
-            services.AddSingleton<IHostedService>(_ => _.GetService<HelloWorldClientHostedService>());
-            services.AddSingleton(_ => _.GetService<HelloWorldClientHostedService>().Client);
-            services.AddControllers();
-        }
-
-        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
-        {
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-            }
-
-            app.UseHttpsRedirection();
-
-            app.UseRouting();
-
-            app.UseAuthorization();
-
-            app.UseEndpoints(endpoints =>
-            {
-                endpoints.MapControllers();
-            });
-        }
-    }
-}
