--- Samples/3.0/HostedService/Controllers/HelloWorldController.cs
@@ -1,13 +1,0 @@
-﻿using Microsoft.AspNetCore.Mvc;
-
-namespace ASPNetCoreHostedServices.Controllers
-{
-    public class HelloWorldController : Controller
-    {
-        // GET
-        public IActionResult Index()
-        {
-            return View();
-        }
-    }
-}
