--- Samples/3.0/HostedService/Orleans/HelloWorldClientHostedService.cs
@@ -1,7 +1,0 @@
-﻿namespace ASPNetCoreHostedServices.Internal
-{
-    public class HelloWorldClientHostedService
-    {
-        
-    }
-}
