--- Samples/3.0/HostedService/Orleans/HelloWorldGrain.cs
@@ -1,16 +1,0 @@
-﻿using System.Threading.Tasks;
-using Microsoft.Extensions.Logging;
-
-namespace ASPNetCoreHostedServices.Internal
-{
-    public class HelloWorldGrain : Orleans.Grain, IHelloWorld
-    {
-        private readonly ILogger<HelloWorldGrain> _logger;
-
-        public HelloWorldGrain(ILogger<HelloWorldGrain> logger) {
-            this._logger = logger;
-        }
-
-        public Task<string> SayHello() => Task.FromResult("Hello world!");
-    }
-}
