--- Samples/3.0/HostedService/Orleans/IHelloWorld.cs
@@ -1,7 +1,0 @@
-﻿namespace ASPNetCoreHostedServices.Orleans
-{
-    public interface IHelloWorld
-    {
-        
-    }
-}
